refactor existing warnings UI
